reorganization for better structure for unit tests
