+++ src/components/LogOutput/LogOutput.vue
