+++ src/views/RunPage.vue
@@ -32,10 +32,11 @@
 
 <script setup lang="ts">
 import { ref } from 'vue';
-import ViewComponent from '@/components/ViewComponent/ViewComponent.vue';
 import { useRunning, isRunning } from '@/composables/useRunning';
-import LogOutput from '@/components/LogOutput.vue';
 import { Vigad } from '@/proc/Vigad';
+
+import ViewComponent from '@/components/ViewComponent/ViewComponent.vue';
+import LogOutput from '@/components/LogOutput/LogOutput.vue';
 
 /**
  * Get singelton instance reference to vigad
